GCPW: Add validation to picture URL
After successful login through GCPW, we download user's profile picture
to be shown in Winlon UI. Project Fortify bug recommends we validate the
picture url before downloading from it.

Typical picture url looks like:
https://lh3.googleusercontent.com/a/ACg8ocK_NQEh4PUNtaO75Cn6ddlzXe_UaAW9uQPSDW5t9R6FAnSj5Q=s96-c

Bug: 498850269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7952050
Commit-Queue: Krishna Kurapati <krkurapati@google.com>
Cr-Commit-Position: refs/heads/main@{#1649143}

+++ chrome/credential_provider/gaiacp/reg_utils.cc
@@ -125,7 +125,7 @@ HRESULT SetMachineRegBinaryInternal(const std::wstring& key_name,
   return S_OK;
 }
 
-std::wstring GetImageRegKeyForSpecificSize(int image_size) {
+std::wstring GetImageRegKeyForSpecificSize(size_t image_size) {
   return kImageRegKey + base::NumberToWString(image_size);
 }
 
@@ -269,7 +269,7 @@ HRESULT GetMachineRegBinaryInternal(const std::wstring& key_name,
 }
 
 HRESULT GetAccountPictureRegString(const std::wstring& user_sid,
-                                   int image_size,
+                                   size_t image_size,
                                    wchar_t* value,
                                    ULONG* length) {
   return GetMachineRegString(GetAccountPictureRegPathForUSer(user_sid),
@@ -279,7 +279,7 @@ HRESULT GetAccountPictureRegString(const std::wstring& user_sid,
 
 // Sets a specific account picture registry key in HKEY_LOCAL_MACHINE
 HRESULT SetAccountPictureRegString(const std::wstring& user_sid,
-                                   int image_size,
+                                   size_t image_size,
                                    const std::wstring& value) {
   return SetMachineRegString(GetAccountPictureRegPathForUSer(user_sid),
                              GetImageRegKeyForSpecificSize(image_size), value);
